# src/utility/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Social Network for Music API",
    "description": "API per il sito Social Network for Music",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/create-playlist": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "summary": "Creazione di una nuova playlist",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titolo": {
                  "example": "any"
                },
                "descrizione": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                },
                "èPrivata": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/update-playlist/{id}": {
      "patch": {
        "tags": [
          "Playlist"
        ],
        "summary": "Modifica di una playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titolo": {
                  "example": "any"
                },
                "descrizione": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                },
                "èPrivata": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-playlists-public": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Restituisce le playlist pubbliche",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-my-playlists": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Restituisce le playlist dell'utente loggato",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-created-playlists": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Restituisce le playlist create dall'utente loggato",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-playlist/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Restituisce le informazioni riguardanti una playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/delete-playlist/{id}": {
      "delete": {
        "tags": [
          "Playlist"
        ],
        "summary": "Elimina una playlist (solo il proprietario)",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add-track-playlist": {
      "post": {
        "tags": [
          "Playlist track "
        ],
        "summary": "Aggiunge una traccia ad una playlist (solo il proprietario)",
        "description": "",
        "parameters": [
          {
            "name": "playlistId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "trackId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/remove-track-playlist": {
      "delete": {
        "tags": [
          "Playlist track"
        ],
        "summary": "Rimuove una traccia da una playlist (solo il proprietario)",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playlistId": {
                  "example": "any"
                },
                "trackId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-playlist-tracks/{id}": {
      "get": {
        "tags": [
          "Playlist track"
        ],
        "summary": "Restituisce le tracce di una playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/follow-playlist/{id}": {
      "post": {
        "tags": [
          "Social network interaction"
        ],
        "summary": "Permette di seguire una playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/unfollow-playlist/{id}": {
      "delete": {
        "tags": [
          "Social network interaction"
        ],
        "summary": "Permette di smettere di seguire una playlist",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}